# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Ops Last Mile Dashboard",
    "version": "0.1.0",
...
  },
  "paths": {
    "/api/management-view/capacity": {
      "get": {
        "summary": "Management View -- Hub Size and Staff headcount per station, from the admin-uploaded Fleet Management workbook. Managers and admins only.",
        "operationId": "capacity_api_management_view_capacity_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapacityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/management-view/notes": {
      "get": {
        "summary": "Management View -- every station's backlog mitigation plan / rescue plan / deployment cost / PTWH count. Managers and admins only.",
        "operationId": "list_notes_api_management_view_notes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StationNote"
                  },
                  "title": "Response List Notes Api Management View Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/management-view/notes/{station_code}": {
      "put": {
        "summary": "Management View -- save one station's backlog mitigation plan / rescue plan / deployment cost / PTWH count. Managers and admins only.",
        "operationId": "put_note_api_management_view_notes__station_code__put",
        "parameters": [
          {
            "name": "station_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationNote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapacityResponse": {
        "properties": {
          "sources": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Sources"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CapacityRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "sources",
          "rows"
        ],
        "title": "CapacityResponse"
      },
      "CapacityRow": {
        "properties": {
          "station_code": {
            "type": "string",
            "title": "Station Code"
          },
          "station_name": {
            "type": "string",
            "title": "Station Name"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqft"
          },
          "staff_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Count"
          },
          "ptwh_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ptwh Count"
          }
        },
        "type": "object",
        "required": [
          "station_code",
          "station_name",
          "zone",
          "region",
          "sqft",
          "staff_count",
          "ptwh_count"
        ],
        "title": "CapacityRow"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteIn": {
        "properties": {
          "mitigation_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitigation Plan"
          },
          "rescue_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rescue Plan"
          },
          "rescue_deployment_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rescue Deployment Cost"
          },
          "ptwh_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ptwh Count"
          }
        },
        "type": "object",
        "title": "NoteIn"
      },
...
      "StationNote": {
        "properties": {
          "station_code": {
            "type": "string",
            "title": "Station Code"
          },
          "mitigation_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitigation Plan"
          },
          "rescue_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rescue Plan"
          },
          "rescue_deployment_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rescue Deployment Cost"
          },
          "ptwh_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ptwh Count"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "station_code",
          "mitigation_plan",
          "rescue_plan",
          "rescue_deployment_cost",
          "ptwh_count",
          "updated_by",
          "updated_at"
        ],
        "title": "StationNote"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}